# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Commodities Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Commodity {
  id: string;
  symbol: string;
  name: string;
  /** metals | energy | agriculture | softs */
  category: string;
  /** e.g. USD/oz, USD/bbl, USD/bu */
  unit: string;
  price: number;
  change: number;
  changePercent: number;
  high24h: number;
  low24h: number;
  lastUpdated: string;
  currency?: string;
  /** @nullable */
  description?: string | null;
}

export interface PricePoint {
  timestamp: string;
  price: number;
}

export interface ProductionRegion {
  id: string;
  commodityId: string;
  name: string;
  country: string;
  lat: number;
  lon: number;
  /** mine | oilfield | farm | reserve | refinery */
  type: string;
  /** Percentage of world production/reserves */
  shareOfWorld: number;
  /** @nullable */
  annualOutput?: string | null;
  /** @nullable */
  outputUnit?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface CommodityDetail {
  id: string;
  symbol: string;
  name: string;
  category: string;
  unit: string;
  price: number;
  change: number;
  changePercent: number;
  high24h: number;
  low24h: number;
  lastUpdated: string;
  currency?: string;
  /** @nullable */
  description?: string | null;
  priceHistory: PricePoint[];
  regions: ProductionRegion[];
}

export interface WeatherForecast {
  date: string;
  maxTemp: number;
  minTemp: number;
  condition: string;
  conditionCode?: number;
  precipitationChance: number;
  windSpeed?: number;
}

export interface WeatherData {
  lat: number;
  lon: number;
  regionName: string;
  temp: number;
  feelsLike?: number;
  humidity: number;
  windSpeed: number;
  condition: string;
  conditionCode: number;
  uvIndex?: number;
  visibility?: number;
  forecast: WeatherForecast[];
}

export interface NewsArticle {
  id: string;
  title: string;
  summary: string;
  source: string;
  publishedAt: string;
  url: string;
  /** Array of commodity IDs this article relates to */
  commodities: string[];
  /** positive | negative | neutral */
  sentiment: string;
  /**
     * regulatory | market | weather | geopolitical | supply
     * @nullable
     */
  category?: string | null;
  /** @nullable */
  imageUrl?: string | null;
}

export interface MarketSummary {
  totalCommodities: number;
  gainersCount: number;
  losersCount: number;
  topGainer: Commodity;
  topLoser: Commodity;
  lastUpdated: string;
  totalMarketChange?: number;
}

export type ListCommoditiesParams = {
/**
 * Filter by category (metals, energy, agriculture)
 */
category?: string;
};

export type ListRegionsParams = {
/**
 * Filter by commodity ID
 */
commodityId?: string;
};

export type GetWeatherParams = {
lat: number;
lon: number;
regionName?: string;
};

export type ListNewsParams = {
/**
 * Filter by commodity symbol (e.g. GOLD, OIL)
 */
commodity?: string;
limit?: number;
};