# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Commodities Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /commodities:
    get:
      operationId: listCommodities
      tags: [commodities]
      summary: List all commodities with live prices
      parameters:
        - in: query
          name: category
          schema:
            type: string
          description: Filter by category (metals, energy, agriculture)
      responses:
        "200":
          description: List of commodities with current prices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Commodity"
...
  /commodities/{symbol}:
    get:
      operationId: getCommodity
      tags: [commodities]
      summary: Get commodity detail with price history
      parameters:
        - in: path
          name: symbol
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Commodity detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommodityDetail"
        "404":
          description: Not found
...
components:
  schemas:
...
    Commodity:
      type: object
      required:
        - id
        - symbol
        - name
        - category
        - unit
        - price
        - change
        - changePercent
        - high24h
        - low24h
        - lastUpdated
      properties:
        id:
          type: string
        symbol:
          type: string
        name:
          type: string
        category:
          type: string
          description: "metals | energy | agriculture | softs"
        unit:
          type: string
          description: "e.g. USD/oz, USD/bbl, USD/bu"
        price:
          type: number
        change:
          type: number
        changePercent:
          type: number
        high24h:
          type: number
        low24h:
          type: number
        lastUpdated:
          type: string
        currency:
          type: string
        description:
          type: ["string", "null"]
    PricePoint:
      type: object
      required:
        - timestamp
        - price
      properties:
        timestamp:
          type: string
        price:
          type: number
    CommodityDetail:
      type: object
      required:
        - id
        - symbol
        - name
        - category
        - unit
        - price
        - change
        - changePercent
        - high24h
        - low24h
        - lastUpdated
        - priceHistory
        - regions
      properties:
        id:
          type: string
        symbol:
          type: string
        name:
          type: string
        category:
          type: string
        unit:
          type: string
        price:
          type: number
        change:
          type: number
        changePercent:
          type: number
        high24h:
          type: number
        low24h:
          type: number
        lastUpdated:
          type: string
        currency:
          type: string
        description:
          type: ["string", "null"]
        priceHistory:
          type: array
          items:
            $ref: "#/components/schemas/PricePoint"
        regions:
          type: array
          items:
            $ref: "#/components/schemas/ProductionRegion"
    ProductionRegion:
      type: object
      required:
        - id
        - commodityId
        - name
        - country
        - lat
        - lon
        - type
        - shareOfWorld
      properties:
        id:
          type: string
        commodityId:
          type: string
        name:
          type: string
        country:
          type: string
        lat:
          type: number
        lon:
          type: number
        type:
          type: string
          description: "mine | oilfield | farm | reserve | refinery"
        shareOfWorld:
          type: number
          description: Percentage of world production/reserves
        annualOutput:
          type: ["string", "null"]
        outputUnit:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
...