# lib/api-zod/src/generated/types/commodity.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Commodities Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Commodity {
  id: string;
  symbol: string;
  name: string;
  /** metals | energy | agriculture | softs */
  category: string;
  /** e.g. USD/oz, USD/bbl, USD/bu */
  unit: string;
  price: number;
  change: number;
  changePercent: number;
  high24h: number;
  low24h: number;
  lastUpdated: string;
  currency?: string;
  /** @nullable */
  description?: string | null;
}