# lib/api-zod/src/generated/types/commodityDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Commodities Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PricePoint } from './pricePoint';
import type { ProductionRegion } from './productionRegion';

export interface CommodityDetail {
  id: string;
  symbol: string;
  name: string;
  category: string;
  unit: string;
  price: number;
  change: number;
  changePercent: number;
  high24h: number;
  low24h: number;
  lastUpdated: string;
  currency?: string;
  /** @nullable */
  description?: string | null;
  priceHistory: PricePoint[];
  regions: ProductionRegion[];
}