# lib/api-zod/src/generated/types/pricePoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Commodities Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface PricePoint {
  timestamp: string;
  price: number;
}