# lib/api-zod/src/generated/types/productionRegion.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Commodities Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductionRegion {
  id: string;
  commodityId: string;
  name: string;
  country: string;
  lat: number;
  lon: number;
  /** mine | oilfield | farm | reserve | refinery */
  type: string;
  /** Percentage of world production/reserves */
  shareOfWorld: number;
  /** @nullable */
  annualOutput?: string | null;
  /** @nullable */
  outputUnit?: string | null;
  /** @nullable */
  description?: string | null;
}